♻️ Apply common project structure
