+++ railway.json
@@ -4,7 +4,7 @@
         "builder": "DOCKERFILE"
     },
     "deploy": {
-        "startCommand": "python main.py",
+        "startCommand": "python -m aicord",
         "restartPolicyType": "ON_FAILURE",
         "restartPolicyMaxRetries": 10
     }
